examples/search_agent/requirements.txt
typer>=0.9.0
rich>=13.0.0
python-dotenv>=1.0.0
